More API implementation

+++ Api/delete_user.php
@@ -20,26 +20,42 @@ if($_SERVER['REQUEST_METHOD']=='POST'){
 
 				if(isset($_POST['user_id'])&& !empty($_POST['user_id'])){
 
-					if(DB::delete_user($conn,$_POST['user_id'])){
+					
+					if(DB::confirm_id($conn,$_POST['user_id'])){
 
-						$data=[
-							'res'=>'User Successfully Deleted',
-							'status'=>200
-						];
+
+						if(DB::delete_user($conn,$_POST['user_id'])){
+
+							$data=[
+								'res'=>'User Successfully Deleted',
+								'status'=>200
+							];
+								
 							
-						
-						echo json_encode($data);
-
+							echo json_encode($data);
+	
+	
+						}else{
+	
+							$data=[
+								'res'=>'Unable To Delete',
+								'status'=>500
+							];
+								
+							
+							echo json_encode($data);
+						}
 
 					}else{
 
 						$data=[
-							'res'=>'Unable To Delete',
-							'status'=>500
+							'res'=>'User not Found',
+							'status'=>404
 						];
 							
 						
 						echo json_encode($data);
+
 					}
 				}else{
